refactor: abstracting logic into seperate files
