+++ interfaces/IAllCoords.ts
@@ -1,0 +1,5 @@
+export default interface IAllCoords {
+  allCoords: number[][];
+  activeCoords: number[][];
+  inactiveCoords: number[][];
+}
